requirements.txt
flask==2.3.3
numpy==1.24.3
matplotlib==3.7.2
Pillow==10.0.0
python-dotenv==1.0.0
werkzeug==2.3.7
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
requests==2.31.0
groq==0.4.0
